Explorer: +resize, +hide
