+++ assets/js/explorerTree/tabHandler.js
@@ -44,13 +44,9 @@ import { ColorComments } from "../editor/colorComments.js"
 
 import { renderPyMsgSuccess, renderPyMsgErr } from "../terminalRenderer/PyRuntimeHandler.js"
 
-import { TypescriptParser } from "../contextParsers/typescriptParser.js"
-import { JavascriptParser } from "../contextParsers/javascriptParser.js"
-import { JSONParser } from "../contextParsers/jsonParser.js"
-import { HTMLParser } from "../contextParsers/htmlParser.js"
-import { CSSParser } from "../contextParsers/cssParser.js"
 import { triggerAceChanged, triggerAceClicked } from "./triggers.js"
 import { TopWindowList, destroyAllTopWindowLists } from "../topWindowHandler/topWindowList.js"
+import { setEditorContext } from "./helpers/setEditorContext.js"
 
 export const recentlyClosed = new Map();
 export const tabsByPath = new Map();
@@ -98,7 +94,6 @@ export class themeEditors {
     }
 
     static add(id, value) {
-        console.log(`Added new theme editor:`, id, `with value:`, value)
         this.themes[id] = value
     }
     static getThemes() {
@@ -618,74 +613,6 @@ export async function openTab(path, content, extension, name, pathContext, isNew
         return;
     }
 
-    const errors = new Map()
-    const markerIds = new Set()
-
-    let diagnosticTimer = null
-    let generation = 0
-    let renderToken = 0
-
-    function addMarker(key, range) {
-        if (markerIds.has(key)) return
-        const id = editor.session.addMarker(range, "error-marker", "fullLine")
-        markerIds.set(key, id)
-    }
-
-    function clearMarkers() {
-        for (const id of markerIds) {
-            editor.session.removeMarker(id)
-        }
-        markerIds.clear()
-
-        editor.renderer.updateFull()
-    }
-
-    function makeErrorKey(item) {
-        return `${item.line}:${path}`
-    }
-
-    function renderErrors() {
-        const token = ++renderToken
-
-        for (const id of markerIds) {
-            editor.session.removeMarker(id)
-        }
-        markerIds.clear()
-
-        const annotations = []
-
-        for (const err of errors.values()) {
-            if (token !== renderToken) return
-
-            err.markerId = null
-            annotations.push(err.annotation)
-
-            const id = editor.session.addMarker(
-                err.range,
-                "error-marker",
-                "fullLine"
-            )
-
-            markerIds.add(id)
-            err.markerId = id
-        }
-
-        editor.session.setAnnotations(annotations)
-        editor.renderer.updateFull()
-    }
-    function clearErrors() {
-        const allMarkers = editor.session.getMarkers()
-        for (const id in allMarkers) {
-            if (allMarkers[id].clazz === "error-marker") {
-                editor.session.removeMarker(Number(id))
-            }
-        }
-        markerIds.clear()
-        errors.clear()
-        editor.session.setAnnotations([])
-        editor.renderer.updateFull()
-    }
-
     function updateEditorData() {
         let cursor = editor.getCursorPosition()
 
@@ -699,142 +626,6 @@ export async function openTab(path, content, extension, name, pathContext, isNew
         setErrors(editor.getSession().getAnnotations())
     }
 
-    async function setEditorContext(properties = {}) {
-        const isErrorsUpdate = properties.errorsUpdate !== false
-
-        updateEditorData()
-
-        clearTimeout(diagnosticTimer)
-        const currentGen = ++generation
-
-        if (language.mode === "javascript") {
-            editor.getSession().setUseWorker(false)
-
-            diagnosticTimer = setTimeout(async () => {
-                const diagnostics = await window.electron.javascriptDiagnostic(editor.getValue())
-
-                if (currentGen !== generation) return
-                if (!isErrorsUpdate) return
-
-                clearErrors()
-
-                diagnostics.forEach(item => {
-                    const range = new aceRange(
-                        item.line - 1,
-                        item.col,
-                        item.line - 1,
-                        999
-                    )
-
-                    errors.set(makeErrorKey(item), {
-                        range,
-                        annotation: {
-                            row: item.line - 1,
-                            column: item.col,
-                            text: item.message,
-                            type: "error"
-                        },
-                        markerId: null
-                    })
-
-                    addRuntimeError({
-                        msg: item.message,
-                        line: item.line,
-                        col: item.col,
-                        time: Math.floor(Date.now() / 1000)
-                    })
-                })
-
-                renderErrors()
-            }, 500)
-
-            const jsParser = new JavascriptParser()
-            const ast = await window.electron.javascriptAST(editor.getValue())
-            const row = editor.getCursorPosition().row + 1
-
-            const chain = jsParser.getContextChain(ast, row)
-            jsParser.renderContext(chain)
-        }
-
-        else if (language.mode === "typescript") {
-            editor.getSession().setUseWorker(false)
-
-            diagnosticTimer = setTimeout(async () => {
-                const diagnostics = await window.electron.typescriptDiagnostic(editor.getValue())
-
-                console.log(diagnostics)
-
-                if (currentGen !== generation) return
-                if (!isErrorsUpdate) return
-
-                clearErrors()
-
-                diagnostics.forEach(item => {
-                    const range = new aceRange(
-                        item.line - 1,
-                        item.col,
-                        item.line - 1,
-                        999
-                    )
-
-                    errors.set(makeErrorKey(item), {
-                        range,
-                        annotation: {
-                            row: item.line - 1,
-                            column: item.col,
-                            text: item.message,
-                            type: "error"
-                        },
-                        markerId: null
-                    })
-
-                    addRuntimeError({
-                        msg: item.message,
-                        line: item.line,
-                        col: item.col,
-                        time: Math.floor(Date.now() / 1000)
-                    })
-                })
-
-                renderErrors()
-            }, 500)
-
-            const tsParser = new TypescriptParser()
-            const ast = await window.electron.typescriptAST(editor.getValue())
-            const row = editor.getCursorPosition().row + 1
-
-            const chain = tsParser.getContextChain(ast, row)
-            tsParser.renderContext(chain)
-        }
-
-        else if (language.mode === "json") {
-            const jsonParser = new JSONParser()
-            jsonParser.showJSONContext(editor, document.querySelector(".code-structure"))
-        }
-
-        else if (language.mode === "html") {
-            const htmlParser = new HTMLParser()
-            htmlParser.showHTMLContext(editor, document.querySelector(".code-structure"))
-        }
-
-        else if (language.mode === "css") {
-            const cssParser = new CSSParser()
-            const row = editor.getCursorPosition().row + 1
-
-            const chain = cssParser.getContextChain(editor.getValue(), row)
-            cssParser.renderContext(chain)
-        }
-
-        else {
-            editor.getSession().setUseWorker(true)
-
-            const codeStructure = document.querySelector(".code-structure")
-            if (codeStructure) {
-                codeStructure.textContent = `Context unavailable for ${language.name}`
-            }
-        }
-    }
-
     ace.config.loadModule(`ace/mode/${language.mode}`, () => {
         if (language.mode === "javascript") {
             initJSSH(editor)
@@ -845,7 +636,12 @@ export async function openTab(path, content, extension, name, pathContext, isNew
     });
 
     editor.session.on('change', async () => {
-        await setEditorContext()
+        await setEditorContext({}, {
+            editor: editor,
+            language: language,
+            updateEditorData: updateEditorData
+        })
+
         triggerAceChanged({ editor: editor, extension: extension, language: language })
 
         // unused find
@@ -864,7 +660,11 @@ export async function openTab(path, content, extension, name, pathContext, isNew
     });
 
     editor.on("click", async () => {
-        await setEditorContext({ errorsUpdate: false })
+        await setEditorContext({ errorsUpdate: false }, {
+            editor: editor,
+            language: language,
+            updateEditorData: updateEditorData
+        })
     });
 
     editor.session.on('changeAnnotation', function () {
