+++ html/index.html
@@ -100,6 +100,7 @@
 					<div class="popup-title" gls="popups.actions.title">Actions</div>
 
 					<div class="popup-content hidden">
+                        <div class="popup-content__divider"></div>
                         <div class="popup-content__item" action="appreload" gls="popups.actions.reload">Reload application</div>
                         <div class="popup-content__item" action="appminimize" gls="popups.actions.minimize">Minimize application</div>
                         <div class="popup-content__item" action="appmaximize" gls="popups.actions.maximize">Maximize application</div>
@@ -284,21 +285,15 @@
             </div>
             <div class="code-footer hidden">
                 <div class="code-footer__section">
-                    <span id="currentLang">Javascript</span>
-                    <span class="code-footer__section-btn" id="code-snippet" tooltip="tooltips.codeSnippet" tooltippos="top"> 
+                    <span class="visible-after-opened-folder" id="currentLang">Javascript</span>
+                    <span class="code-footer__section-btn visible-after-opened-folder" id="code-snippet" tooltip="tooltips.codeSnippet" tooltippos="top"> 
                         <span class="material-symbols-rounded">camera</span>
                     </span>
-                    <span class="code-footer__section-btn" id="code-copy" tooltip="tooltips.copy" tooltippos="top">
+                    <span class="code-footer__section-btn visible-after-opened-folder" id="code-copy" tooltip="tooltips.copy" tooltippos="top">
                         <span class="material-symbols-rounded">content_copy</span>
                     </span>
-                    <span class="code-footer__section-btn" id="code-console" tooltip="tooltips.terminal" tooltippos="top">
+                    <span class="code-footer__section-btn visible-after-opened-folder" id="code-console" tooltip="tooltips.terminal" tooltippos="top">
                         <span class="material-symbols-rounded">terminal_2</span>
-                    </span>
-                    <span class="code-footer__section-btn" id="js-minify" visibleOn="language:javascript" tooltip="tooltips.jsMinify" tooltippos="top">
-                        <span class="material-symbols-rounded">compress</span>
-                    </span>
-                    <span class="code-footer__section-btn" id="css-minify" visibleOn="language:css" tooltip="tooltips.cssMinify" tooltippos="top">
-                        <span class="material-symbols-rounded">compress</span>
                     </span>
                     <span class="code-footer__section-btn row hidden" id="incontext-loading">
                         <l-ring
@@ -311,18 +306,23 @@
                     </span>
                 </div>
                 <div class="code-footer__section">
-                    <span class="code-footer__section-btn" id="changeTabSize">Tab size: <span id="currentTabSize">0</span></span>
-                    <span><span id="currentSymbols">0</span> Symb.</span>
-                    <span><span id="currentCol">0</span> Col</span>
-                    <span>Line <span id="currentLine">0</span></span>
-                    <span class="text-danger text-in-line" id="bottomErrors">
+                    <span class="code-footer__section-btn visible-after-opened-folder" id="changeTabSize">Tab size: <span id="currentTabSize">0</span></span>
+                    <span class="visible-after-opened-folder"><span id="currentSymbols">0</span> Symb.</span>
+                    <span class="visible-after-opened-folder"><span id="currentCol">0</span> Col</span>
+                    <span class="visible-after-opened-folder">Line <span id="currentLine">0</span></span>
+                    <span class="text-danger text-in-line visible-after-opened-folder" id="bottomErrors">
                         <span class="material-symbols-rounded">cancel</span>
                         <span id="errorCount">0</span>
                     </span>
-                    <span class="text-warning text-in-line" id="bottomWarnings">
+                    <span class="text-warning text-in-line visible-after-opened-folder" id="bottomWarnings">
                         <span class="material-symbols-rounded">error</span>
                         <span id="warningCount">0</span>
                     </span>
+                    <div class="left-divider">
+                        <span class="code-footer__section-btn active" id="sidebar-toggle" tooltip="tooltips.sidebar" tooltippos="top">
+                            <span class="material-symbols-rounded">dock_to_right</span>
+                        </span> 
+                    </div>
                 </div>
             </div>
         </div>
