+++ package.json
@@ -7,7 +7,8 @@
     "scripts": {
         "build": "tsc",
         "start": "npm run build && electron .",
-        "dist": "npm run build && electron-builder"
+        "dist": "npm run build && electron-builder",
+        "dist-linux": "npm run build && electron-builder --linux"
     },
     "build": {
         "appId": "com.codemotion",
